requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-crispy-forms==1.9.1
django-filter==2.2.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.1.0
yapf==0.30.0
